+++ CPP_09/ex01/RPN.hpp
@@ -12,7 +12,7 @@ using std::stack;
 
 class RNP {
 private:
-    stack<int> operands;
+    stack<float> operands;
 
 public:
     RNP();
